requirements.txt
Flask
cryptography
pandas
requests
python-dotenv
